requirements.txt
Django~=3.2.9
django-taggit==2.0.0
Pillow